+++ src/codesyncd/populate_buffer.ts
@@ -13,13 +13,14 @@ import {pathUtils} from "../utils/path_utils";
 import {
     formatDatetime,
     getBranch,
-    getSkipPaths,
+    getDefaultIgnorePatterns,
+    getGlobIgnorePatterns,
     getSyncIgnoreItems,
     isEmpty,
-    isIgnoreAblePath,
     isUserActive,
     readFile,
-    readYML
+    readYML,
+    shouldIgnorePath
 } from "../utils/common";
 import {
     GLOB_TIME_TAKEN_THRESHOLD, 
@@ -122,6 +123,7 @@ class PopulateBuffer {
     deletedRepoBranchPath: string;
     originalsRepoBranchPath: string;
     syncIgnoreItems: string[];
+    defaultIgnorePatterns: string[];
     potentialMatchFiles: string[];
     gotPotentialMatchFiles: boolean;
 
@@ -144,6 +146,7 @@ class PopulateBuffer {
         this.deletedRepoBranchPath = this.pathUtils.getDeletedRepoBranchPath();
         this.originalsRepoBranchPath = this.pathUtils.getOriginalsRepoBranchPath();
         this.syncIgnoreItems = getSyncIgnoreItems(this.repoPath);
+        this.defaultIgnorePatterns = getDefaultIgnorePatterns();
         this.potentialMatchFiles = [];
         this.gotPotentialMatchFiles = false;
     }
@@ -224,13 +227,13 @@ class PopulateBuffer {
         - Shadow file should not be a binary file since we are going to match the text of files
         - Shadow file should not be empty
          */
-        const skipPaths = getSkipPaths(this.shadowRepoBranchPath, this.syncIgnoreItems);
+        const globIgnorePatterns = getGlobIgnorePatterns(this.shadowRepoBranchPath, this.syncIgnoreItems);
         // These are shadow files whose actual files are present in the repo
         const skipShadowRelPaths = this.itemPaths.map(itemPath => itemPath.rel_path);
         const t0 = new Date().getTime();
         const shadowRelPaths = globSync("**", { 
             cwd: this.shadowRepoBranchPath,
-            ignore: skipPaths, 
+            ignore: globIgnorePatterns, 
             nodir: true, 
             dot: true
         });
@@ -242,9 +245,13 @@ class PopulateBuffer {
         const filteredFiles = shadowFilePaths.filter(shadowFilePath => {
             const relPath = shadowFilePath.split(path.join(this.shadowRepoBranchPath, path.sep))[1];
             const isInConfig = relPath in this.configFiles;
-            const shouldIgnorePath = isIgnoreAblePath(relPath, this.syncIgnoreItems);
+            if (!isInConfig) {
+                removeFile(shadowFilePath, "getPotentialRenamedFiles");
+                return false;
+            }
+            const ignoreAblePath = shouldIgnorePath(relPath, this.defaultIgnorePatterns, this.syncIgnoreItems);
             // If file is not in config OR is is present in .syncignore, remove the file from .shadow
-            if (!isInConfig || shouldIgnorePath) {
+            if (ignoreAblePath) {
                 removeFile(shadowFilePath, "getPotentialRenamedFiles");
                 return false;
             }
@@ -309,7 +316,8 @@ class PopulateBuffer {
             const cacheFilePath = path.join(this.deletedRepoBranchPath, relPath);
             const shadowFilePath = path.join(this.shadowRepoBranchPath, relPath);
             // See if should discard this file
-            if (isIgnoreAblePath(relPath, this.syncIgnoreItems) ||
+            const isIgnoreablePath = shouldIgnorePath(relPath, this.defaultIgnorePatterns, this.syncIgnoreItems);
+            if (isIgnoreablePath  ||
                 activeRelPaths.includes(relPath) ||
                 this.renamedFiles.includes(relPath) ||
                 fs.existsSync(cacheFilePath) ||