+++ CHANGELOG.md
@@ -4,9 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [3.21.0] - 2023-05-16 [NOT DEPLOYED]
+## [3.21.0] - 2023-05-22
 - Using fs.unlink instead of fs.unlinkSync
 - Running 1 instance of populateBuffer, waiting before previous iteration is completed
+- Using .syncignore on system level to ignore default directories/files like .git, node_modules etc
 
 ## [3.20.2] - 2023-05-16
 - Fixed skip patterns for Windows